# Conflicts:
#	SP4/Base.v12.suo
#	SP4/Base/Debug/Base.log
#	SP4/Common/Debug/Common.log

+++ SP4/Base/Source/Main/Game/Objects/Characters/CharacterEntity.cpp
@@ -52,7 +52,7 @@ Skill* CharacterEntity::GetSkillInVector(string SkillName)
 
 void CharacterEntity::SetDamageMitigation()
 {
-	this->DamageMitigation = (0.06 * Defence) / (1 + 0.06 * Defence);
+	this->DamageMitigation = (int)((0.06f * Defence) / (1 + 0.06f * Defence));
 }
 
 void CharacterEntity::Init(int Level)
